+++ website/src/components/MatchScouting/AutoScoringComponents/AutoPositionSelector.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
-import FullFieldMap from "../../assets/FullFieldMap.png";
+import FullFieldMap from "../../../assets/FullFieldMap.png";
 import { toast } from "react-toastify";
 
 import {
