+++ website/src/components/MatchScouting/AutoScoringComponents/pathUtils.js
