+++ website/src/pages/MatchScouting/GameStartPage.jsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import { useLocation } from "react-router-dom";
 
 import ProceedBackButton from "../../components/ProceedBackButton";
-import GameStartSelectAlliance from "../../components/GameStartComponents/GameStartSelectAlliance";
-import GameStartTextInput from "../../components/GameStartComponents/GameStartTextInput";
-import GameStartSelectTeam from "../../components/GameStartComponents/GameStartSelectTeam";
+import GameStartSelectAlliance from "../../components/MatchScouting/GameStartComponents/GameStartSelectAlliance";
+import GameStartTextInput from "../../components/MatchScouting/GameStartComponents/GameStartTextInput";
+import GameStartSelectTeam from "../../components/MatchScouting/GameStartComponents/GameStartSelectTeam";
 
 /**
  * Renders a component representing the Game Start Page.
